[IntersectionObserver] Fix cross-origin check for honoring rootMargin
An issue on the IntersectionObserver spec repo raised the issue of
the precise limitation on rootMargin and rootBounds. The original spec
language refers to the "unit of related similar-origin browsing
contexts", which suggests that cross-origin same-site observations
could use rootMargin and rootBounds. That is incorrect; same-site
cross-origin observations should *not* be able to use rootMargin and
rootBounds.

Chromium already conforms to the more restrictive behavior; this CL
is just to add a test to enforce it.

https://github.com/w3c/IntersectionObserver/pull/448

Separately, the existing cross-origin-iframe.sub.html test was disabled
due to flakiness; and then a subsequent code change broke the
cross-origin limitation for applying rootMargin, so the test has been
reliably failing. This patch fixes the cross-origin check and reenables
the test.

Change-Id: Ib864f9a9433b4d323ef73a3b2b5a8f645697c07c