Support drag and drop in Devtool Input Protocol
When we are doing a mouse move with left button pressed, we should start
a drag gesture, and  generate events like "dragstart", "dragenter",
"dragover", "dragleave" and "dragend". Because drag events are generated
from OS system, we have to manually call the drag functions in
RenderWidgetHostImpl from devtool input handler to generate all the drag
events.

Bug: 850071
Change-Id: I3e1bf2b183bf1eb9bad78fa3ee594e757a1e2bab