Default disable scroll animations in web tests
Our end goal is to enable routing input through the compositor input
handler, even in single thread mode (the default in web tests).
Scroll animations make this difficult because this mode doesn't have a
scheduler; there's nothing to tick new animation frames. Main thread
animations work by calling ScheduleAnimationForWebTests on the
WebWidgetTestProxy to get a new frame. Trying a similar approach from
the compositor thread led to visual changes in test output and proved
difficult to make work (changed the timing of when pixel dumps were
taken).

Instead, this CL effectively disables scroll animations (by default) in
web tests. This should have the additional benefit of making tests less
flaky and run faster. It will also make it clear that web tests
shouldn't be used to test animation behavior since tests can't reliably
guarantee frame timing.

In the compositor, animations are disabled by overriding the animation
curve duration to a near 0 value. This ensures the animation is
completed in one tick and code that creates an animation already
schedules a new frame due to needing a commit. We do this, rather than
using a non-animated path, to ensure code coverage remains close to real
world.

For consistency, we also turn off the main thread scroll animator. It
would be nice to do this by shortening the curve duration there too but
this mechanism doesn't yet exist on the main thread but it does have a
straightforward way to disable the scroll animator so we use that.

There are some tests, primarily those in fast/scroll-behavior, that were
trying to test animation behavior. These are largely flaky and
ineffective; some of these are being rewritten as browser tests in
crbug.com/1047176. However, we put these into a virtual suite that
enables animations to preserve current behavior.

Bug: 915926
Change-Id: Id4716628837385bb91f682049369b96c9d8b3833