[cleanup] Remove intrinsic overrides in LayoutImage, etc.
Initially added in:
https://chromium-review.googlesource.com/c/chromium/src/+/1150725/
https://chromium-review.googlesource.com/c/chromium/src/+/1191926/
https://chromium-review.googlesource.com/c/chromium/src/+/1210544/

This now appears to be (somewhat) dead code. Previous patches removed
the intrinsicSize attribute which this was initially hooked up to.

This code actually slightly dangerous at the moment, as it appears that
if you have certain feature policies turned on you will get subtly
different behaviour that what you expected.

E.g. the "document-write" policy?
https://source.chromium.org/chromium/chromium/src/+/master:third_party/blink/renderer/core/feature_policy/feature_policy_features.json5;l=132?q=experimentalproductivityfeatures%20file:json5&ss=chromium%2Fchromium%2Fsrc&originalUrl=https:%2F%2Fcs.chromium.org%2F

This leaves the reporting behaviour in:
https://source.chromium.org/chromium/chromium/src/+/master:third_party/blink/renderer/core/layout/layout_image.cc;l=437?q=isdefaultintrinsicsize&ss=chromium&originalUrl=https:%2F%2Fcs.chromium.org%2F

Change-Id: I80d1d376753070dc94559e3177d31c0fd3439700