[css-grid] Use MarginAfter with flipped blocks writing mode
When computing the logical offset of non-static absolute positioned
grid items, we were using the child's MarginBefore logical function to
get the margin based on its writing-mode.

Although we are looking for the left/top margin to compute the offset,
when the item has flipped block writing mode (vertical-rl) we need to
switch to the right/bottom margin, hence we'll call MarginAfter in this
case.

This changes also performs some adjustments in RTL cases, since the
logical position of an absolute positioned item is computed differently
for vertical-rl containers.

Bug: 1156225
Change-Id: I03796f65178f931df11c16d4187c6708f758d1b1