[css-flex] Ignore items' heights for intrinsic block size calculation
Blink currently makes this flex item 200px tall. Firefox gives it the
correct height of 100px.

<div style="display: flex; flex-direction: column">
  <div style="flex-basis: 100px; height: 200px">
    <div style="height: 100%">

Blink was letting the inner div's height resolve against the flex item's
200px, giving the flex item an automatic minimum size of 200px, even
though the min-content size is supposed to be calculated with the item
having an 'auto' size. Meaning we should ignore the item's 200px height.

To keep the 200px from having an effect, this patch broadens the
definition of ConstraintSpace().IsFixedBlockSizeDefinite. It no longer
has to accompany a FixedBlockSize. It now means that the element is to
be laid out as if it has an auto height.

Flexbox now sets this constraint on items when the flexbox needs the
item's intrinsic block size.

Fixed: 1033476, 607585
Change-Id: I0376e8ec71ef5a6a71989a8bdda63b39718852e8