[selectors] :focus-visible on <select> element
When you do a mouse click to focus a <select> element
:focus-visible shouldn't match.
This is Firefox behavior and has been agreed on the following issue:
https://github.com/w3c/csswg-drafts/issues/5822

This patch fixes the problem by removing
HTMLSelectElement::MayTriggerVirtualKeyboard() that was returning true
for <select> elements.
And also by passing the proper FocusType::kMouse
in MenuListSelectType::DefaultEventHandler().

The previous wrong behavior was tested in focus-visible-002.html,
so removed the <select> element on that test.

Added a new test case to check the new behavior
in focus-visible-003.html and focus-visible-004.html.

BUG=1169708
TEST=external/wpt/css/selectors/focus-visible-003.html
TEST=external/wpt/css/selectors/focus-visible-004.html

Change-Id: I6649ce6355aafc79b2fa3d258b2b69fa5a259ce7