Implement fall-back logic for composite bgcolor animation
We can't handle all background color animations on the compositor
thread, therefore we need a fallback logic that we can still animate
them on the main thread.

Specifically, we'd only animate a background color animation if:
1. It is the one with highest composite order.
2. It is composite:replace.

In this CL though, we made it simpler. We would ensure that there
is only one bgcolor animation on the element, and that animation
must be composite:replace, in order to composite it.

Bug: 1153672
Change-Id: I1362038b2d6e78f640df0aac52914d58c200589f