Enable subframe navigation to urn:uuid resources in Web Bundles
This implements subframe loading with Web Bundles. Design doc:
https://docs.google.com/document/d/1_AqUBS4Gr45MPPtXGTUl7Q0DMIEw2zUeWO0zo6Sj0z4/edit#heading=h.aay60dv2vzua

blink::FrameLoader::StartNavigation attaches a WebBundleToken to the
request. It is propagated to the (new) web_bundle_token field of
mojom::BeginNavigationParams, and used to construct a navigation request
in NavigationURLLoaderImpl.

Created subframes have opaque-origins, and are hosted by a process for
the Bundle's origin. To achieve these, this CL:

- Adds a field for WebBundle URL to content::UrlInfo
- For navigation to a urn: URL,
  - SiteInstanceImpl::GetSiteForURLInternal() uses the Bundle URL
  - NavigationRequest::GetOriginForURLLoaderFactory() returns an opaque
    origin derived from the bundle's origin

Bug: 1082020
Change-Id: I07e03cce9f7da7eb6fd9d4218cf31882585b8463