Stop preloading subresources which should be loaded from WebBundles.
Currently subresources for static elements are not correctly loaded from
WebBundles (<link rel=webbundle>). This is because the PreloadScanner
creates PreloadRequests for the static elements without knowing the
WebBundle links <link rel=webbundle>, and the HTMLDocumentParser starts
the preload requests via HTMLResourcePreloader without associating the
requests with the WebBundles.

To fix this issue, this CL changes the PreloadScanner not to create the
PreloadRequests for the subresources which should be loaded from the
WebBundles by introducing the PreloadRequest::ExclusionInfo class.

Bug: 1122514
Change-Id: I87a0be177408585d4f815a36940980b859f5d970