Do not use plugin-types for trusted-types WPTs
Some web platform tests for trusted types trigger a violation of the
CSP directive plugin-types to verify that a CSP violation trigger. The
plugin-types directive is specified with an invalid (hence in practice
empty) value. With the removal of the 'plugin-types' directive
implementations, these tests will break.

This CL converts those tests to trigger violations of the 'object-src'
directive instead.

Bug: 1168001
Change-Id: I042a7c600bc5c7caeb6dd87ff94b181b5fb66b6b