Change whitespace collapsing in Mozilla counter style WPT tests so that Blink can pass
In Firefox, white spaces in list marker text are never collapsed [1],
which can be a either bug or just a UA-defined behavior. The Mozzila
WPT tests currently assume this behavior. Blink doesn't have this
behavior, and therefore is currently failing on these tests.

For now, we modify the tests by adding 'white-space: pre' to list
markers, so that Blink can also pass them.

[1] https://bugzilla.mozilla.org/show_bug.cgi?id=1688769

Bug: 687225
Change-Id: I5b1e244a963e7454231d2d674950485fc8a06ffd